inheritable version-dependent QueryInfo
